+++ types/utils/isObject.d.ts
@@ -1,8 +1,8 @@
-declare const _default: (value: any) => boolean;
 /**
  * Check if param is an object.
  *
  * @param {any} value Character strings, arrays, objects, etc. to be checked.
  * @returns {boolean} Returns true if it is an object.
  */
+declare const _default: (value: any) => boolean;
 export default _default;