+++ types/utils/isString.d.ts
@@ -1,8 +1,8 @@
-declare const _default: (value: any) => boolean;
 /**
   * Return true if `val` is a non-empty string.
   *
   * @param  {any} `val` The value to check
   * @return {boolean}
   */
+declare const _default: (value: any) => boolean;
 export default _default;